+++ src/Exception/ConnectTimeoutException.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GuzzleHttp\Exception;
+
+/**
+ * Exception thrown when connection establishment exceeds the time limit.
+ */
+class ConnectTimeoutException extends ConnectException
+{
+}