+++ src/Exception/NetworkTimeoutException.php
@@ -5,7 +5,8 @@ declare(strict_types=1);
 namespace GuzzleHttp\Exception;
 
 /**
- * Exception thrown when a transfer times out before a response is received.
+ * Exception thrown when a transfer times out before response headers are
+ * received.
  */
 class NetworkTimeoutException extends NetworkException
 {