+++ src/Exception/ResponseTimeoutException.php
@@ -5,7 +5,8 @@ declare(strict_types=1);
 namespace GuzzleHttp\Exception;
 
 /**
- * Exception thrown when a transfer times out after a response is received.
+ * Exception thrown when a transfer times out after response headers are
+ * received.
  */
 class ResponseTimeoutException extends ResponseException
 {