+++ src/Exception/TooManyRedirectsException.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace GuzzleHttp\Exception;
 
+/**
+ * Exception thrown when redirect middleware reaches the redirect limit.
+ */
 class TooManyRedirectsException extends ResponseException
 {
 }